backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pyjwt==2.8.1
motor==3.3.2
pymongo==4.6.0
httpx==0.25.2
requests==2.31.0
bcrypt==4.1.1
python-multipart==0.0.6
cors==1.0.1
